+++ Nyanble/Features/Recommendations/Intent/IntentPerformer.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+protocol IntentPerformer {
+    associatedtype Input
+    associatedtype Output
+    static func perform(_ input: Input) async throws -> Output
+}
