+++ Nyanble/Features/Recommendations/Model/AIRecommendations.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+@Generable
+struct AIRecommendations {
+    @Guide(description: "A list of recommended nearby places", .count(3))
+    var places: [AIRecommendedPlace]
+}
