+++ Nyanble/Features/Recommendations/Model/AIRecommendedPlace.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+@Generable
+struct AIRecommendedPlace {
+    var name: String
+    var detail: String
+}
