+++ Nyanble/Features/Recommendations/Model/RecommendedPlace.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+struct RecommendedPlace: Identifiable, Hashable {
+    let id = UUID()
+    let name: String
+    let detail: String
+}
